Add RTTY functionality and refactor related components
- Introduced ScreenAMRTTY for RTTY decoding display and handling.
- Updated configuration structure to include rttyShiftFrequency and rttyBaudRate.
- Refactored existing code to replace rttyShiftHz with rttyShiftFrequency.
- Enhanced UI components to support RTTY settings and display.
- Updated debug output to reflect new RTTY parameters.

+++ include/decoder_api.h
@@ -187,5 +187,5 @@ struct DecodedData {
     // RTTY-specifikus státuszok (Core1 írja, Core0 olvassa)
     volatile uint16_t rttyMarkFreq;  // Mark frekvencia (Hz)
     volatile uint16_t rttySpaceFreq; // Space frekvencia (Hz)
-    volatile uint16_t rttyBaudRate;  // Baud sebesség (beállított)
+    volatile float rttyBaudRate;     // Baud sebesség (beállított)
 };
